Minimum required
The minimum amount of code required to add a working Debug Screen to your mod. Copy the code in ITMPlugin.HandleInput() and ITMPlugin.Draw(), as well as the variables defined at the top of the class to your mod to add a working Debug Screen. From there, you can start adding values to debug.